pyproject.toml
[tool.poetry]
name = "jlsa"
version = "0.1.0"
description = ""
authors = ["kuznetsow42 <kuznetsow42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
django-cors-headers = "^4.6.0"
django-rest-durin = "^1.1.0"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
